Fix drawing broken when canvas is drawn across multiple windows
This is an insane edge case, but it happened to me when I tried rendering a barcode in a popup window using bwip-js by creating a div in one window, appending it to another, and using React portal to render the UI including the barcode. When I did that, canvas was no longer an instance of HTMLCanvasElement apparently, and as a result, rendering was simply aborted.

This fixes that issue.